separation of logic for the meta controller
