+++ ExpenseTracker/Controllers/MetaController.cs
@@ -1,32 +1,20 @@
-using ExpenseTracker.Core.Enums;
+using ExpenseTracker.Services;
 using Microsoft.AspNetCore.Mvc;
-using System.ComponentModel.DataAnnotations;
-using System.Reflection;
 
 namespace ExpenseTracker.Controllers;
 
 [ApiController]
 [Route("api/v1/[controller]")]
-public class MetaController : ControllerBase
+public class MetaController(IMetaService metaService) : ControllerBase
 {
+	private readonly IMetaService _metaService = metaService;
+
 	[HttpGet("get-expense-categories")]
 	public IActionResult GetExpenseCategories()
 	{
-		var categories = Enum.GetValues(typeof(ExpenseCategory))
-			.Cast<ExpenseCategory>()
-			.Select(c => new
-			{
-				value = c.ToString(),
-				displayName = GetDisplayName(c)
-			});
+		var categories = _metaService.GetExpenseCategories()
+			.Select(c => new { c.value, c.displayName });
 
 		return Ok(categories);
-	}
-
-	private static string GetDisplayName(Enum value)
-	{
-		return value.GetType()
-			.GetMember(value.ToString())[0]
-			.GetCustomAttribute<DisplayAttribute>()?.Name ?? value.ToString();
 	}
 }
