+++ ExpenseTracker/Services/IMetaService.cs
@@ -1,0 +1,6 @@
+namespace ExpenseTracker.Services;
+
+public interface IMetaService
+{
+	IEnumerable<(string value, string displayName)> GetExpenseCategories();
+}
